sqld,wal_hook: port to using ffi libsql_wal_methods
Previous implementation of the WAL hook was coded before official
ffi bindings were introduced. Now, the implementation of WalMethodsHook
embeds the libsql_wal_methods struct, so that it's guaranteed to be 100%
compatible with the C API.
Also, one crucial fix was added to allow nesting virtual WAL implementations:
xOpen now calls the original xOpen with a pointer to the underlying methods,
so the underlying virtual WAL also has a chance to initialize itself.