replica: ensure page headers store pages in frame order
This commit makes sure that page headers are added in the dirty list in the
order of arrival. A database corruption was observed after a sufficiently large
transaction actually had two versions of a single page in a batch. Frames are
applied in dirty-list order, so due to the fact that we previously appended new
frames to the beginning on the list, we ended up writing the newest version of
the page first, and then rewriting it with the old one, hence the
inconsistency. Whew.